+++ domain-api/src/main/kotlin/de/chrgroth/spotify/control/domain/port/in/HealthPort.kt
@@ -4,5 +4,4 @@ import de.chrgroth.spotify.control.domain.model.HealthStats
 
 interface HealthPort {
     fun getStats(): HealthStats
-    fun activatePartition(partitionKey: String): Boolean
 }