+++ domain-api/src/main/kotlin/de/chrgroth/spotify/control/domain/port/in/PlaybackPort.kt
@@ -1,7 +1,6 @@
 package de.chrgroth.spotify.control.domain.port.`in`
 
 import arrow.core.Either
-import de.chrgroth.outbox.OutboxTaskResult
 import de.chrgroth.spotify.control.domain.error.DomainError
 import de.chrgroth.spotify.control.domain.model.UserId
 import de.chrgroth.spotify.control.domain.outbox.DomainOutboxEvent
@@ -15,8 +14,8 @@ interface PlaybackPort {
     fun rebuildPlaybackData(userId: UserId)
     fun appendPlaybackData(userId: UserId)
     fun syncArtistPlaybackFromPlaylists(userId: UserId)
-    fun handle(event: DomainOutboxEvent.FetchCurrentlyPlaying): OutboxTaskResult
-    fun handle(event: DomainOutboxEvent.FetchRecentlyPlayed): OutboxTaskResult
-    fun handle(event: DomainOutboxEvent.RebuildPlaybackData): OutboxTaskResult
-    fun handle(event: DomainOutboxEvent.AppendPlaybackData): OutboxTaskResult
+    fun handle(event: DomainOutboxEvent.FetchCurrentlyPlaying): Either<DomainError, Unit>
+    fun handle(event: DomainOutboxEvent.FetchRecentlyPlayed): Either<DomainError, Unit>
+    fun handle(event: DomainOutboxEvent.RebuildPlaybackData): Either<DomainError, Unit>
+    fun handle(event: DomainOutboxEvent.AppendPlaybackData): Either<DomainError, Unit>
 }