+++ domain-api/src/main/kotlin/de/chrgroth/spotify/control/domain/port/in/PlaylistCheckPort.kt
@@ -1,9 +1,10 @@
 package de.chrgroth.spotify.control.domain.port.`in`
 
-import de.chrgroth.outbox.OutboxTaskResult
+import arrow.core.Either
+import de.chrgroth.spotify.control.domain.error.DomainError
 import de.chrgroth.spotify.control.domain.outbox.DomainOutboxEvent
 
 interface PlaylistCheckPort {
-    fun handle(event: DomainOutboxEvent.RunPlaylistChecks): OutboxTaskResult
-    fun getDisplayNames(): Map<String, String>
+  fun handle(event: DomainOutboxEvent.RunPlaylistChecks): Either<DomainError, Unit>
+  fun getDisplayNames(): Map<String, String>
 }