+++ domain-impl/src/test/kotlin/de/chrgroth/spotify/control/domain/PlaybackEnrichmentAdapterTests.kt
@@ -10,7 +10,6 @@ import de.chrgroth.spotify.control.domain.port.out.AppPlaybackRepositoryPort
 import de.chrgroth.spotify.control.domain.port.out.AppPlaylistCheckRepositoryPort
 import de.chrgroth.spotify.control.domain.port.out.AppTrackRepositoryPort
 import de.chrgroth.spotify.control.domain.port.out.DashboardRefreshPort
-import de.chrgroth.spotify.control.domain.port.out.OutboxManagementPort
 import de.chrgroth.spotify.control.domain.port.out.OutboxPort
 import de.chrgroth.spotify.control.domain.port.out.PlaylistRepositoryPort
 import de.chrgroth.spotify.control.domain.port.out.SpotifyAccessTokenPort
@@ -33,7 +32,6 @@ class PlaybackEnrichmentAdapterTests {
     private val appPlaybackRepository: AppPlaybackRepositoryPort = mockk(relaxed = true)
     private val userRepository: UserRepositoryPort = mockk(relaxed = true)
     private val outboxPort: OutboxPort = mockk()
-    private val outboxManagement: OutboxManagementPort = mockk()
     private val playlistRepository: PlaylistRepositoryPort = mockk()
     private val playlistCheckRepository: AppPlaylistCheckRepositoryPort = mockk()
     private val dashboardRefresh: DashboardRefreshPort = mockk(relaxed = true)
@@ -48,7 +46,6 @@ class PlaybackEnrichmentAdapterTests {
         appPlaybackRepository,
         userRepository,
         outboxPort,
-        outboxManagement,
         playlistRepository,
         playlistCheckRepository,
         dashboardRefresh,